Remove the world-specific skin folders
* First, that feature introduced crashes on clients in cases where there
integrated server hasn't been initialized yet, i.e. the user directly
went to multiplayer without opening a singleplayer world first.
* Second, it was only rarely used. In an ideal world, all data related
to a world would be stored in that world. But most mods don't allow for
such a thing anyway, so you store all your extra-world-data in another
folder elsewhere and then copy or link the required files over when you
need to. And, on another note ... who thought about what happens if a
skin is required for two separate worlds?